requirements.txt
google-generativeai
dotenv
flask